Not a super well tested commit here. Planning to rework a large section of DialogESD.cs and I want to commit all the working stuff first. This commit includes a vareity of additions to papyrus filters and calls. It also implements factions partially. Additionally, the players race has now been implemented, and in order to do that we had to add the hks extension plugin. So when building for this version you need to add that to your dll chain. This commit also implements dead/disable flags and spawn management scripts. It also adds deadtype count for papyrus filters. Things that die are now dead and no longer respawn. I also added a save data reset button to the church of elleh for debugging. That will reset all data to default for the morrowind world. This is useful for testing. Prolly some other stuff but that's all I can remember!

+++ JortPob/Common/Utility.cs
@@ -93,7 +93,7 @@ namespace JortPob.Common
             fmg.Entries = fmg.Entries.AsParallel().OrderBy(entry => entry.ID).ToList();
         }
 
-        public static int Pow(int x, uint pow)
+        public static long Pow(int x, uint pow)
         {
             int ret = 1;
             while (pow != 0)
